refactor: created two different functions for ICP and for ETH to decode the QR code
